Fix 15 issues from the v1.12 feature review round
High:
- Wipe compare rendered raw uint16 frames as RGB888 (corrupted output)
  for 16-bit stacks; display conversion now applied
- ECC on machines with CuPy truncated aligned 16-bit frames to uint8
  (mod-256 wrap), silently corrupting results and poisoning the disk
  cache; GPU path preserves the source dtype
- Rotate/flip/resize served stale pre-transform results from the
  registration disk cache; the signature now includes the frame shape
  (also covers load-time downsampling) and superseded cache folders are
  pruned automatically

Medium:
- Registration cache signature missed the load-time downsample scale
- The silent update check wrote its daily rate-limit timestamp only on
  states it never reached, so it contacted GitHub at every launch
- EXIF orientation was skipped on the drag-drop and batch preload paths
- CLI batch silently overwrote stacks with duplicate folder basenames
- Quick-preview registration-only view showed garbage for 16-bit

Low:
- Output-list icons/thumbnails corrupted for 16-bit results
- Removed a shadowed duplicate load_from_folder definition
- CLI: warns when --output is combined with --output-dir and reports
  skipped non-folder inputs
- QActionGroup leaked per context-menu open
- Label colours scale into the 16-bit domain when annotating uint16
  frames (were drawn near-black)

+++ core/registration.py
@@ -690,7 +690,7 @@ def _align_ecc_impl(input_source, output_path=None, img_filenames=None, downscal
             aligned_img_gpu = cp.stack(channels, axis=2)
             
             # 传回 CPU
-            aligned_img = cp.asnumpy(aligned_img_gpu).astype(np.uint8)
+            aligned_img = cp.asnumpy(aligned_img_gpu).astype(img.dtype)
             
         else:
             # CPU 版本 (OpenCV)
